backend/requirements.txt
audioread==2.1.8
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
decorator==4.4.1
Flask==1.1.1
Flask-CLI==0.4.0
Flask-SQLAlchemy==2.4.1
google-api-core==1.16.0
google-auth==1.10.1
google-cloud==0.34.0
google-cloud-core==1.2.0
google-cloud-speech==1.3.1
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
librosa==0.7.2
llvmlite==0.31.0
MarkupSafe==1.1.1
nltk==3.4.5
numba==0.47.0
numpy==1.18.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
python-dotenv==0.10.4
pytz==2019.3
requests==2.22.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.12
urllib3==1.25.7
Werkzeug==0.16.0
wincertstore==0.2
